// src/models/ConfigModel.tsx
export interface ConfigModel {
    canBeNull: boolean
    cate: string
    enable: boolean
    isSecret: boolean
    key: string
    max: number
    maxLength: number
    min: number
    name: string
    nullTips: string
    placeholder: string
    type: string
    value: string
    dependKey: string
    attribute: any
}